server/requirements.txt
 ## Install Packages

# Django Stuff
django==2.2.10
django-extensions==3.1.3
django-environ==0.4.5
djangorestframework==3.11.1
django-searchable-encrypted-fields==0.1.9
pycryptodome==3.10.1
django-cors-headers==3.7.0
psycopg2==2.8.6

# Dev Stuff
ipdb==0.13.9
ipython==7.24.1

# Python Stuff
requests==2.25.1
urllib3==1.26.5
python-dateutil==2.8.1

# Celery Stuff
celery==4.4.7
django-celery-beat==2.0.0
django-celery-results==1.2.1
redis==3.5.3

